+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Add `async` option to `start_link`
+
 ## 0.1.3
 - `WebSockex.start_link` will no longer cause the calling process to exit on
   connection failure and will return a proper error tuple instead.